+++ src/graphics/render_address.rs
@@ -5,7 +5,7 @@ use crate::utils::BitGroup;
 ///
 /// It's a 15-bit address used for both reading and writing PPU memory through
 /// PPUDATA ($2007) register
-#[derive(Copy, Clone)]
+#[derive(Copy, Clone, Default)]
 pub struct RenderAddress {
     value: BitGroup<u16>,
 }