+++ src/main.rs
@@ -6,6 +6,7 @@ fn main() {
     env_logger::init();
 
     let mut nes = Nes::default();
+    nes.setup_tv();
     nes.connect_controller_one(ControllerButtons::default());
     // let cartidge = Cartidge::new(Path::new("/path/to/cartidge"));
     let cartidge = Cartidge::new("roms/Super Mario Bros. (World).nes");