+++ src/types.rs
@@ -7,10 +7,13 @@ use crate::interfaces::Memory;
 use crate::processor::bus::Bus;
 use crate::processor::memory::{Ciram, MirroredMemory, Ram, Rom};
 
-pub type SharedPpu = Rc<RefCell<Ppu>>;
 pub type SharedBus = Rc<RefCell<Bus>>;
+
 pub type SharedMemory = Rc<RefCell<dyn Memory>>;
 pub type SharedRam = Rc<RefCell<Ram>>;
 pub type SharedCiram = Rc<RefCell<Ciram>>;
-pub type SharedController = Rc<RefCell<Controller>>;
 pub type SharedMirroredRom = Rc<RefCell<MirroredMemory<Rom>>>;
+
+pub type SharedPpu = Rc<RefCell<Ppu>>;
+
+pub type SharedController = Rc<RefCell<Controller>>;